Remove error return statements if the nsq message is invalid
If the nsq message publish is invalid, returning an error will requeue
the mesage back into nsq where it will be retried 5 times. Since the
message is invalid, there isn't really any point in retrying since it
will be invalid for the next 5 times so it is better to log the error
and return nil.