Consolidate the handoff_snapshot_table migrations into one file.
When creating a new database, I would be missing a few of the fields and
we don't need the history of how to alter the table when we can just
have a migration to create the table that we want. NOTE: I only changed
this because my dump.sql after running migrations from scratch was
removing a few columns.